+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file
 - add TUFX compat
 - add ClickThroughBlocker compat
 - update netkan config
+- force disable physics range extender
 
 
 ## 0.0.5 - 2025-07-28